requirements.txt
tls-client
fake-useragent
colorama
python-dotenv
ollama